refactor(workflow): move movie workflow to workflows/movie subpackage; combine transcode+move; add retries
- Move all movie.*.go files from workflows/ to workflows/movie/ (package movie)
- Combine transcode and move into a single step: runTranscode now writes directly
  to outputDir/._<basename>.tmp and atomically renames to the final path, eliminating
  the cross-filesystem copy and the temp-file-gone crash window
- Add WithRetries(3) to lookup, notify-radarr, and cleanup tasks
- Extract shared test helpers (copyTestVideo, testVideoPath) to testhelpers_test.go
- Update cmd/worker to import workflows/movie
