+++ workflows/movie/cleanup_test.go
@@ -1,4 +1,4 @@
-package workflows
+package movie
 
 import (
 	"os"
