+++ workflows/movie/probe.go
@@ -1,4 +1,4 @@
-package workflows
+package movie
 
 import (
 	"context"
