+++ workflows/movie/probe_test.go
@@ -1,4 +1,4 @@
-package workflows
+package movie
 
 import (
 	"os"
@@ -8,21 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-// testVideoPath points to the small H.264/MP4 clip shared with the ffprobe package.
-const testVideoPath = "../pkg/ffprobe/testdata/video.mp4"
-
-// copyTestVideo copies the shared test video to a temp file and returns its path.
-// The file is NOT registered for cleanup so tests can verify deletion behaviour.
-func copyTestVideo(t *testing.T) string {
-	t.Helper()
-	src, err := os.ReadFile(testVideoPath)
-	require.NoError(t, err)
-
-	dst := filepath.Join(t.TempDir(), "video.mp4")
-	require.NoError(t, os.WriteFile(dst, src, 0o600))
-	return dst
-}
 
 func TestRunProbe(t *testing.T) {
 	tests := []struct {
