+++ workflows/movie/transcode.go
@@ -1,0 +1,63 @@
+package movie
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/solidDoWant/media-processor/pkg/ffmpeg"
+	"github.com/solidDoWant/media-processor/pkg/ffprobe"
+)
+
+// selectVideoCodec returns CodecCopy when the video is already H.264 or H.265 in an
+// MKV container, and CodecH265 otherwise.
+func selectVideoCodec(videoCodecName, format string) ffmpeg.Codec {
+	if strings.Contains(format, string(ffmpeg.ContainerMKV)) {
+		if videoCodecName == ffprobe.CodecNameH264 || videoCodecName == ffprobe.CodecNameH265 {
+			return ffmpeg.CodecCopy
+		}
+	}
+
+	return ffmpeg.CodecH265
+}
+
+// runTranscode transcodes input.FilePath into outputDir, writing to a temp file named
+// "._<basename>.tmp" and atomically renaming it to "<basename>" on success.
+// Writing directly to the output directory avoids a cross-filesystem copy and
+// guarantees the rename is atomic on Linux (same directory).
+func runTranscode(ctx context.Context, input MovieInput, probe probeOutput, outputDir string) error {
+	videoCodec := selectVideoCodec(probe.VideoCodec, probe.Format)
+
+	baseName := filepath.Base(input.FilePath)
+	tempPath := filepath.Join(outputDir, "._"+baseName+".tmp")
+	finalPath := filepath.Join(outputDir, baseName)
+
+	if err := ffmpeg.NewTranscode(input.FilePath, tempPath).
+		ToVideoCodec(videoCodec).
+		ToContainer(ffmpeg.ContainerMKV).
+		Build().
+		Run(ctx); err != nil {
+		if removeErr := os.Remove(tempPath); removeErr != nil && !errors.Is(removeErr, os.ErrNotExist) {
+			return errors.Join(
+				fmt.Errorf("transcode: %w", err),
+				fmt.Errorf("cleanup temp file: %w", removeErr),
+			)
+		}
+		return fmt.Errorf("transcode: %w", err)
+	}
+
+	if err := os.Rename(tempPath, finalPath); err != nil {
+		if removeErr := os.Remove(tempPath); removeErr != nil && !errors.Is(removeErr, os.ErrNotExist) {
+			return errors.Join(
+				fmt.Errorf("move output file: %w", err),
+				fmt.Errorf("cleanup temp file: %w", removeErr),
+			)
+		}
+		return fmt.Errorf("move output file: %w", err)
+	}
+
+	return nil
+}
